+++ src/music_flac/api/__init__.py
@@ -1,5 +1,6 @@
 from music_flac.api.base import FlacSource
+from music_flac.api.hifi_flac import HifiFlacSource
 from music_flac.api.http import HttpFlacSource
 from music_flac.api.stub import StubFlacSource
 
-__all__ = ["FlacSource", "HttpFlacSource", "StubFlacSource"]
+__all__ = ["FlacSource", "HifiFlacSource", "HttpFlacSource", "StubFlacSource"]